Rename stroke recognition files
