+++ include/utils/StrokeRecognition.h
