+++ meson.build
@@ -24,7 +24,7 @@ src = [
     'src/utils/misc.c',
     'src/utils/eta-keyboard.cpp',
     'src/utils/Storage.cpp',
-    'src/utils/stroke_recognition.cpp',
+    'src/utils/StrokeRecognition.cpp',
 ]
 
 version = '0.0.0'
